Split world rendering helper structs into a header
