+++ src/render/DeferredShading.cpp
@@ -8,6 +8,7 @@
 #include "../scene/Camera.hpp"
 #include "../scene/Light.hpp"
 #include "../scene/World.hpp"
+#include "../scene/WorldRenderStructs.hpp"
 #include "../utils/Profiler.hpp"
 #include "../utils/Ui.hpp"
 #include "../utils/Utils.hpp"
