+++ src/render/ForwardRenderer.cpp
@@ -9,6 +9,7 @@
 #include "../scene/Mesh.hpp"
 #include "../scene/Scene.hpp"
 #include "../scene/World.hpp"
+#include "../scene/WorldRenderStructs.hpp"
 #include "../utils/Profiler.hpp"
 #include "../utils/Ui.hpp"
 #include "../utils/Utils.hpp"
