+++ src/render/GBufferRenderer.cpp
@@ -8,6 +8,7 @@
 #include "../scene/Mesh.hpp"
 #include "../scene/Scene.hpp"
 #include "../scene/World.hpp"
+#include "../scene/WorldRenderStructs.hpp"
 #include "../utils/Profiler.hpp"
 #include "../utils/Utils.hpp"
 #include "LightClustering.hpp"
