+++ src/render/LightClustering.cpp
@@ -7,6 +7,7 @@
 #include "../scene/Camera.hpp"
 #include "../scene/Light.hpp"
 #include "../scene/World.hpp"
+#include "../scene/WorldRenderStructs.hpp"
 #include "../utils/Profiler.hpp"
 #include "RenderResources.hpp"
