+++ src/render/SkyboxRenderer.cpp
@@ -4,6 +4,7 @@
 #include "../gfx/VkUtils.hpp"
 #include "../scene/Camera.hpp"
 #include "../scene/World.hpp"
+#include "../scene/WorldRenderStructs.hpp"
 #include "../utils/Profiler.hpp"
 #include "../utils/Utils.hpp"
 #include "RenderResources.hpp"
