+++ src/render/rtdi/RtDiSpatialReuse.cpp
@@ -3,6 +3,7 @@
 #include "../../scene/Camera.hpp"
 #include "../../scene/Light.hpp"
 #include "../../scene/World.hpp"
+#include "../../scene/WorldRenderStructs.hpp"
 #include "../../utils/Profiler.hpp"
 #include "../../utils/Utils.hpp"
 #include "../GBufferRenderer.hpp"
