+++ src/render/rtdi/RtDiTrace.cpp
@@ -5,6 +5,7 @@
 #include "../../scene/Camera.hpp"
 #include "../../scene/Scene.hpp"
 #include "../../scene/World.hpp"
+#include "../../scene/WorldRenderStructs.hpp"
 #include "../../utils/Profiler.hpp"
 #include "../../utils/Ui.hpp"
 #include "../../utils/Utils.hpp"
