+++ src/scene/World.hpp
@@ -2,57 +2,13 @@
 #define PROSPER_SCENE_WORLD_HPP
 
 #include "../gfx/Fwd.hpp"
-#include "../scene/Texture.hpp"
 #include "../utils/Fwd.hpp"
-#include "../utils/Utils.hpp"
 #include "Fwd.hpp"
 
 #include <filesystem>
 #include <memory>
 #include <wheels/allocators/scoped_scratch.hpp>
 #include <wheels/containers/static_array.hpp>
-
-struct WorldDSLayouts
-{
-    uint32_t materialSamplerCount{0};
-    vk::DescriptorSetLayout materialDatas;
-    vk::DescriptorSetLayout materialTextures;
-    vk::DescriptorSetLayout geometry;
-    vk::DescriptorSetLayout modelInstances;
-    vk::DescriptorSetLayout rayTracing;
-    vk::DescriptorSetLayout lights;
-    vk::DescriptorSetLayout skybox;
-};
-
-struct WorldByteOffsets
-{
-    uint32_t modelInstanceTransforms{0};
-    uint32_t previousModelInstanceTransforms{0};
-    uint32_t directionalLight{0};
-    uint32_t pointLights{0};
-    uint32_t spotLights{0};
-    uint32_t globalMaterialConstants{0};
-};
-
-struct WorldDescriptorSets
-{
-    vk::DescriptorSet lights;
-    wheels::StaticArray<vk::DescriptorSet, MAX_FRAMES_IN_FLIGHT> materialDatas;
-    vk::DescriptorSet materialTextures;
-    vk::DescriptorSet geometry;
-    vk::DescriptorSet skybox;
-};
-
-struct SkyboxResources
-{
-    TextureCubemap texture;
-    Image irradiance;
-    Image specularBrdfLut;
-    Image radiance;
-    wheels::Array<vk::ImageView> radianceViews;
-    Buffer vertexBuffer;
-    vk::Sampler sampler;
-};
 
 class World
 {
